requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Redis (async support needed for infra WebSocket/pubsub layer)
redis[asyncio]>=5.0.0

# AWS S3
boto3>=1.34.0

# Environment variables
python-dotenv>=1.0.0

# Data validation (used by FastAPI models)
pydantic>=2.0.0

# AI Integrations
Pillow>=10.0.0         # text slide PNG generation
numpy>=1.26.0          # audio mixing
google-genai>=0.5.0    # Vertex Imagen 3 / Lyria 2
google-cloud-aiplatform>=1.50.0  # Vertex AI SDK
google-cloud-texttospeech>=2.16.0  # Google Cloud TTS
runwayml>=0.3.0        # Runway Gen-4 video
elevenlabs>=1.0.0      # ElevenLabs music/TTS
ideogram-ai>=0.1.0     # Ideogram image generation (via httpx if not available)

# Testing
pytest>=8.0.0
httpx>=0.27.0          # required by FastAPI TestClient and AI integrations
